Add install_audit_permission command. refactoring
