+++ django_auditmatic/install/audit_permission.py
@@ -1,0 +1,7 @@
+
+def install_audit_permission():
+    """
+        install audit permission for configured models.
+    :return:
+    """
+    raise NotImplementedError
